More flexible pelican config/run.
The current buildsite.py process assumes the content is placed into
"content". This is fixed through a new CONTENT_DIR variable.

generate_settings() makes some assumptions about the location of the
theme and any plugins. Clarify this with SETTINGS_DIR variable.

For now, just hack in a look for a "site" directory, which we knows
contains all of the above. If a "content" directory is present, then
it is the original layout, and we keep on rollin'

Added some docco at the head to specify what is needed to run this
script in a local dev/test scenario.