local buildsite script simplified
local buildsite script now just does :tm:

if no pelican-local pipenv is detected in `~` one will be created.
if a website directory is found in the pipenv, local changes take
precedence and no pull operation is performed.

Once the website is built, it will be served locally.

On a re-run, the generated site content and all build artifacts are
deleted before the site is re-built and served.