Restrict jinja version to 3.0.3
This is the most conservative change that fixes #29:
because pelican 4.6.0 specifies `jinja2>=2.7`, but
jinja2 versions of 3.1.0 or later conflict with
pelican 4.6.0, this leads to a broken environment.

This conflict is fixed on pelican 4.7.0 or later,
so when pelican is upgraded this workaround can be
removed, but updating pelican might need further testing.